tried to make more complete and easy to explain components
